get last post by user, posts from last x days

+++ post-service/src/main/java/com/pintertamas/befake/postservice/controller/PostController.java
@@ -1,11 +1,12 @@
 package com.pintertamas.befake.postservice.controller;
 
+import com.amazonaws.services.drs.model.AccessDeniedException;
 import com.amazonaws.services.mq.model.NotFoundException;
+import com.pintertamas.befake.postservice.exception.WrongFormatException;
 import com.pintertamas.befake.postservice.model.Post;
 import com.pintertamas.befake.postservice.service.JwtUtil;
 import com.pintertamas.befake.postservice.service.PostService;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.tomcat.util.http.fileupload.FileUploadException;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -39,6 +40,9 @@ public class PostController {
             Long userId = jwtUtil.getUserIdFromToken(headers);
             Post post = postService.createPost(userId, main, selfie, location);
             return new ResponseEntity<>(post, HttpStatus.CREATED);
+        } catch (WrongFormatException e) {
+            log.error(e.getMessage());
+            return new ResponseEntity<>("Wrong format", HttpStatus.BAD_REQUEST);
         } catch (IOException e) {
             log.error(e.getMessage());
             return new ResponseEntity<>("Could not upload image", HttpStatus.INTERNAL_SERVER_ERROR);
@@ -72,29 +76,68 @@ public class PostController {
     }
 
     @PatchMapping("/{postId}")
-    public ResponseEntity<?> addDescription(@PathVariable Long postId, @RequestParam String description) {
+    public ResponseEntity<?> addDescription(
+            @PathVariable Long postId,
+            @RequestParam String description,
+            @RequestHeader HttpHeaders headers) {
         try {
+            if (jwtUtil.isNotPostOwner(headers, postId))
+                throw new AccessDeniedException("You are not the owner of this post");
             postService.addDescription(postId, description);
             return new ResponseEntity<>(description, HttpStatus.OK);
         } catch (NotFoundException e) {
-            return new ResponseEntity<>("Could not find post", HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
+        } catch (AccessDeniedException e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.UNAUTHORIZED);
         } catch (Exception e) {
             return new ResponseEntity<>("Could not add description", HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @DeleteMapping("/{postId}")
-    public ResponseEntity<String> deletePost(@PathVariable Long postId) {
+    public ResponseEntity<String> deletePost(
+            @PathVariable Long postId,
+            @RequestHeader HttpHeaders headers) {
         try {
+            if (jwtUtil.isNotPostOwner(headers, postId))
+                throw new AccessDeniedException("You are not the owner of this post");
             postService.deletePost(postId);
             return new ResponseEntity<>(postId + " successfully deleted", HttpStatus.OK);
         } catch (NotFoundException e) {
-            return new ResponseEntity<>("Could not find post", HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
         } catch (IllegalArgumentException e) {
-            return new ResponseEntity<>("Could not delete post", HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
+        } catch (AccessDeniedException e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.UNAUTHORIZED);
         } catch (Exception e) {
             return new ResponseEntity<>("Could not delete post", HttpStatus.INTERNAL_SERVER_ERROR);
         }
+    }
 
+    @GetMapping("/lastPosts/{userId}/{daysAgo}")
+    public ResponseEntity<?> getLastPostsFromUser(
+            @PathVariable Long userId,
+            @PathVariable int daysAgo) {
+        try {
+            List<Post> posts = postService.getPostsFromLastXDays(userId, daysAgo);
+            return new ResponseEntity<>(posts, HttpStatus.OK);
+        } catch (NotFoundException e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
+        } catch (Exception e) {
+            return new ResponseEntity<>("Could not query posts", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @GetMapping("/today/{userId}")
+    public ResponseEntity<?> getLastPostBy(
+            @PathVariable Long userId) {
+        try {
+            Post post = postService.getLastPostBy(userId);
+            return new ResponseEntity<>(post, HttpStatus.OK);
+        } catch (NotFoundException e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
+        } catch (Exception e) {
+            return new ResponseEntity<>("Could not query post", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 }
